Preliminary notes on how to fire up an entirely OS X native environment instead of using Vagrant.
No, I'm not religious on this point. Frankly, I just wanted to see if it could be done. For the most part, it looks like it can, and with nothing more bizarre than Homebrew and RVM.

Three specs currently fail that pass under vagrant. All are part of the PrettyText tests, and appear to be whitespace differences, possibly caused by slightly different gem versions.  I have not investigated deeper, yet.